refactor: separate useRouter
