+++ src/main.js
@@ -1,91 +1,6 @@
-import LoginPage from "./pages/LoginPage";
-import MainPage from "./pages/MainPage";
-import NotFoundPage from "./pages/NotFoundPage";
-import ProfilePage from "./pages/ProfilePage";
-import { useUserStore } from "./stores/useUserStore";
+import { useRouter } from "./utils/useRouter";
 
-const routes = {
-  "/": () => renderMainPage(),
-  "/login": () => renderLoginPage(),
-  "/profile": () => renderProfilePage(),
-};
-
-const router = (path = window.location.pathname || "/") => {
-  const route = routes[path];
-  const isLogin = useUserStore.isLogin();
-
-  if (route) {
-    if (path === "/profile" && !isLogin) {
-      history.pushState(null, "", "/login");
-      return router("/login");
-    } else if (path === "/login" && isLogin) {
-      history.pushState(null, "", "/");
-      return router("/");
-    }
-
-    history.pushState(null, "", path);
-    route();
-  } else {
-    document.getElementById("root").innerHTML = `${NotFoundPage()}`;
-  }
-};
-
-const renderMainPage = () => {
-  document.getElementById("root").innerHTML = `${MainPage()}`;
-  document
-    .querySelector("nav")
-    .addEventListener("click", (e) => clickEventHandler(e));
-};
-
-const renderLoginPage = () => {
-  document.getElementById("root").innerHTML = `${LoginPage()}`;
-
-  const loginForm = document.querySelector("#login-form");
-  loginForm.addEventListener("submit", (e) =>
-    submitEventHandler({ e, onSuccess: () => router("/") }),
-  );
-};
-
-const renderProfilePage = () => {
-  document.getElementById("root").innerHTML = `${ProfilePage()}`;
-  document
-    .querySelector("nav")
-    .addEventListener("click", (e) => clickEventHandler(e));
-
-  const profileForm = document.querySelector("#profile-form");
-  profileForm.addEventListener("submit", (e) => submitEventHandler({ e }));
-};
-
-const clickEventHandler = (e) => {
-  const { id, href, tagName } = e.target;
-
-  if (tagName === "A") {
-    e.preventDefault();
-
-    let path = href.slice(href.lastIndexOf("/"));
-
-    if (id === "logout") {
-      useUserStore.resetUserInfoInLocalStorage();
-      path = "/login";
-    }
-
-    router(path);
-  }
-};
-
-const submitEventHandler = ({ e, onSuccess }) => {
-  e.preventDefault();
-  const formData = new FormData(e.target);
-  const userInfo = {
-    username: formData.get("username"),
-    email: formData.get("email") ?? "",
-    bio: formData.get("bio") ?? "",
-  };
-
-  useUserStore.setUserInfoInLocalStorage(userInfo);
-
-  onSuccess && onSuccess();
-};
+const { router } = useRouter();
 
 window.addEventListener("load", () => router());
 window.addEventListener("popstate", () => router());
